refactor(#4159): split the scan gap-recording out of scanner.rs (campsite rule)
scanner.rs went 473 -> 833 (limit 800) on this branch, so this diff is what
pushed it over — the case the campsite rule exists for (epic #1116). Same
pattern already used here for refusal.rs, discovery_walk.rs and
header_minimal.rs.

Moved to the new cqlite-core/src/discovery/scan_gap.rs (120 lines): the
UnreadableDirectory type and the five filesystem-read/gap-recording helpers
read_dir_named, dir_entry_named, entry_is_dir, note_unreadable, kind_of.
scanner.rs (744) keeps the Scanner entry point, the keyspace/table
enumeration, ScanResult/KeyspaceInfo/TableInfo and
has_cassandra_table_uuid_suffix.

PURE MOVE, verified rather than asserted: all five function bodies and the
struct are byte-identical to their pre-split text (brace-matched extraction
compared against git HEAD). The only edits the move forced are `fn` ->
`pub(super) fn`, the import line, and re-exporting UnreadableDirectory from
scan_gap in discovery/mod.rs. No behaviour, signature or public-path change
— `cqlite_core::discovery::UnreadableDirectory` resolves as before.

FILE-SIZE OPT-OUT for the remaining four, acknowledged rather than widened
(#1116 source splits / #1135 test splits):

  cqlite-core/src/error.rs             738 -> 864  (+126)
  cqlite-core/src/storage/mod.rs       798 -> 812  (+14)
  cqlite-core/src/storage/sstable/mod.rs  2793 -> 2806  (+13)
  cqlite-core/src/storage/sstable/refresh.rs  803 -> 827  (+24)

All four were ALREADY over threshold before this branch, and the deltas are
small. error.rs is +126 because of one new enum variant and its docs; an
error enum does not split meaningfully. Splitting a 2806-line module as part
of this diff is exactly the widening being avoided on a release-critical
path. Re-run with CQLITE_ALLOW_FILE_GROWTH=1, which surfaces as
`file-size: OPT-OUT (ratchet NOT enforced)` in the summary rather than a
silent PASS.
