+++ .gate-lite-r1.txt.heartbeat
@@ -1,0 +1,13 @@
+==== AGENT-GATE HEARTBEAT ====
+run-id: /tmp/agent-gate.UUUM0K
+mode: lite
+gate-pid: 399976
+host: ip-172-31-6-169
+beater-pid: 400030
+parent-check: starttime
+interval: 20
+beat-seq: 79
+beat-epoch: 1788819387
+beat-utc: 2026-09-07T22:16:27Z
+logs: /tmp/agent-gate.UUUM0K
+==== END AGENT-GATE HEARTBEAT ====
