+++ cqlite-core/src/discovery/scanner.rs
@@ -41,6 +41,9 @@
 
 use std::path::{Path, PathBuf};
 
+use super::scan_gap::{
+    dir_entry_named, entry_is_dir, kind_of, note_unreadable, read_dir_named, UnreadableDirectory,
+};
 use crate::error::{Error, Result};
 
 /// Keyspace information
@@ -65,24 +68,6 @@ pub struct TableInfo {
     pub sstable_count: usize,
     /// Table directory path
     pub path: PathBuf,
-}
-
-/// One directory a scan could not read.
-///
-/// Structured rather than a `warnings` string: a caller has to be able to FAIL
-/// CLOSED on this (and to name the path to an operator) without parsing prose, and
-/// the [`std::io::ErrorKind`] is what separates a permissions problem from a
-/// directory that was removed mid-scan.
-#[derive(Debug, Clone)]
-pub struct UnreadableDirectory {
-    /// The directory that could not be read.
-    pub path: PathBuf,
-    /// What it was being read as ("keyspace", "table", …).
-    pub role: String,
-    /// The original failure's kind.
-    pub kind: std::io::ErrorKind,
-    /// The original failure, rendered.
-    pub message: String,
 }
 
 /// Result of scanning a data directory
@@ -124,80 +109,6 @@ fn has_cassandra_table_uuid_suffix(dir_name: &str) -> bool {
     } else {
         false
     }
-}
-
-/// Read `dir`, naming it (and what it was being read AS) on failure.
-///
-/// Issue #4159: a directory this scanner cannot read must not read as an EMPTY
-/// directory — an unreadable keyspace and a keyspace with no tables are different
-/// facts, and a caller cannot tell them apart from a short result set.
-fn read_dir_named(dir: &Path, what: &str) -> Result<std::fs::ReadDir> {
-    std::fs::read_dir(dir).map_err(|e| {
-        Error::Io(std::io::Error::new(
-            e.kind(),
-            format!("Failed to read {what} directory {}: {e}", dir.display()),
-        ))
-    })
-}
-
-/// Unwrap one `ReadDir` item, naming the directory it came from on failure.
-///
-/// `ReadDir` yields `io::Result<DirEntry>`: an entry can fail MID-ITERATION (a
-/// concurrent unlink, an I/O fault). `Iterator::flatten` discards those silently,
-/// which is how an entry came to "never exist".
-fn dir_entry_named(
-    item: std::io::Result<std::fs::DirEntry>,
-    dir: &Path,
-) -> Result<std::fs::DirEntry> {
-    item.map_err(|e| {
-        Error::Io(std::io::Error::new(
-            e.kind(),
-            format!("Failed to read an entry of {}: {e}", dir.display()),
-        ))
-    })
-}
-
-/// Record `e` against `dir` as a gap, rather than swallowing it or aborting.
-fn note_unreadable(
-    into: &mut Vec<UnreadableDirectory>,
-    dir: &Path,
-    role: &str,
-    e: &Error,
-    kind: std::io::ErrorKind,
-) {
-    tracing::warn!(
-        "SSTable discovery could not read {role} directory {}: {e}. The scan is \
-         INCOMPLETE.",
-        dir.display()
-    );
-    into.push(UnreadableDirectory {
-        path: dir.to_path_buf(),
-        role: role.to_string(),
-        kind,
-        message: e.to_string(),
-    });
-}
-
-/// The [`std::io::ErrorKind`] behind a crate [`Error`], for the gap record.
-fn kind_of(e: &Error) -> std::io::ErrorKind {
-    match e {
-        Error::Io(io) => io.kind(),
-        _ => std::io::ErrorKind::Other,
-    }
-}
-
-/// Is `entry` a directory? Propagates the `stat` failure instead of answering
-/// `false`, which is what `Path::is_dir()` does for a directory it cannot stat.
-fn entry_is_dir(entry: &std::fs::DirEntry) -> Result<bool> {
-    entry.file_type().map(|ft| ft.is_dir()).map_err(|e| {
-        Error::Io(std::io::Error::new(
-            e.kind(),
-            format!(
-                "Failed to determine the file type of {}: {e}",
-                entry.path().display()
-            ),
-        ))
-    })
 }
 
 /// Scanner for discovering SSTables in a data directory
